\name{pheno.miRNA}
\alias{pheno.miRNA}
\docType{data}
\title{
Phenotypical miRNA information
}
\description{Phenotypical miRNA information
}
\usage{data(pheno.miRNA)}
\format{
  A data frame with 12 observations on the following 2 variables.
  \describe{
    \item{\code{group}}{a factor with levels \code{H} (Healthy) \code{D} (Disease)}
    \item{\code{DvH}}{a numeric vector}
  }
}
\source{
Modified from: \emph{Affò S, Dominguez M, Lozano JJ, Sancho-Bru P et al. Transcriptome analysis identifies TNF superfamily receptors as potential therapeutic targets in alcoholic hepatitis. Gut 2013 Mar;62(3):452-60. PMID: 22637703}
}
\references{
Affò S, Dominguez M, Lozano JJ, Sancho-Bru P et al. Transcriptome analysis identifies TNF superfamily receptors as potential therapeutic targets in alcoholic hepatitis. Gut 2013 Mar;62(3):452-60. PMID: 22637703
}
\examples{
data(pheno.miRNA)
pheno.miRNA
}
\keyword{datasets}
